Refactoring testing tools into test directory
This removes an eslint warning, that those modules were using
packages not part of the final package.
